pay_service/requirements.txt
# fastapi==0.111.1
# pydantic==2.8.2
# pydantic-settings==2.3.4
# geopy==2.4.1
# uvicorn==0.30.3

aio-pika==9.4.2
boto3==1.34.154
python-dotenv==1.0.1